requirements.txt
# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1

# Speech Recognition
whisper>=1.1.10
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0

# NLP and Grammar Analysis
language-tool-python>=2.7.1
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
